+++ src/pages/SearchResultPage.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from "react";
 import "../styles/SearchResultPage.css";
 import { useLocation } from "react-router-dom";
-import useSearchResults from "../components/UseSearchResults";
-
+import useSearchResults from "../components/hooks/UseSearchResult";
 import NavigationContainer from "../containers/NavigationBarContainer";
 import CategoryBarContainer from "../containers/CategoryBarContainer";
 import ProductList from "../components/ProductList";
 import Footer from "../components/Footer";
+import { fetchBasicProductsList } from '../utils/ApiClient';
 
 function SearchResultPage({ setSearchTerm: initialSetSearchTerm }) {
   const location = useLocation();
@@ -31,15 +31,33 @@ function SearchResultPage({ setSearchTerm: initialSetSearchTerm }) {
     isLowFat: false,
   });
 
+  const [productsResult, setProductsResult] = useState({ items: [], error: null, noProductsFound: false });
+
   useEffect(() => {
     setSearchTerm(initialSearchTerm);
   }, [location]);
 
+  useEffect(() => {
+    const fetchData = async () => {
+      const searchParam = searchTerm || initialSearchTerm;
+      console.log("Fetching results for:", searchParam);
+
+      const result = await fetchBasicProductsList(searchParam, selectedCategories, filters, kkiniGreenCheck);
+      setProductsResult(result);
+    };
+
+    if (searchTerm || initialSearchTerm) {
+      fetchData();
+    }
+  }, [searchTerm, selectedCategories, filters, kkiniGreenCheck, initialSearchTerm]);
+
   const { categoryGroups, loading, error, noProductsFound } = useSearchResults(
-    searchTerm,
-    selectedCategories,
-    filters,
-    kkiniGreenCheck
+      searchTerm,
+      selectedCategories,
+      filters,
+      kkiniGreenCheck,
+      initialSearchTerm,
+      productsResult
   );
 
   const handleFilterChange = (updatedFilters) => {
@@ -51,34 +69,34 @@ function SearchResultPage({ setSearchTerm: initialSetSearchTerm }) {
   };
 
   return (
-    <div className="search-result-page">
-      <NavigationContainer
-        searchTerm={searchTerm}
-        setSearchTerm={setSearchTerm}
-        autocompleteItems={autocompleteItems}
-        setAutocompleteItems={setAutocompleteItems}
-      />
-      <div className="page-tit content-max">
-        <h1>"{searchTerm}"에 대한 검색결과</h1>
-      </div>
-      <div className="product-wrap content-max">
-        <CategoryBarContainer
-          onKkiniChecked={handleKkiniGreenCheckChange}
-          onCategoryChange={setSelectedCategories}
-          onFilterChange={handleFilterChange}
-          searchTerm={searchTerm}
-          filters={filters}
-          kkiniGreenCheck={kkiniGreenCheck}
+      <div className="search-result-page">
+        <NavigationContainer
+            searchTerm={searchTerm}
+            setSearchTerm={setSearchTerm}
+            autocompleteItems={autocompleteItems}
+            setAutocompleteItems={setAutocompleteItems}
         />
-        <div className="product-list-wrapper">
-          <ProductList
-            categoryGroups={categoryGroups}
-            noProductsFound={noProductsFound}
-          />
+        <div className="page-tit content-max">
+          <h1>"{searchTerm}"에 대한 검색결과</h1>
         </div>
+        <div className="product-wrap content-max">
+          <CategoryBarContainer
+              onKkiniChecked={handleKkiniGreenCheckChange}
+              onCategoryChange={setSelectedCategories}
+              onFilterChange={handleFilterChange}
+              searchTerm={searchTerm}
+              filters={filters}
+              kkiniGreenCheck={kkiniGreenCheck}
+          />
+          <div className="product-list-wrapper">
+            <ProductList
+                categoryGroups={categoryGroups}
+                noProductsFound={noProductsFound}
+            />
+          </div>
+        </div>
+        <Footer />
       </div>
-      <Footer />
-    </div>
   );
 }
